requirements.txt
# Core dependencies
numpy==1.24.3
pandas==2.0.3
scipy==1.10.1

# Machine Learning
scikit-learn==1.3.0
xgboost==1.7.6
lightgbm==4.0.0
catboost==1.2.1

# Visualization
matplotlib==3.7.2
seaborn==0.12.2
plotly==5.15.0

# Web Application
streamlit==1.25.0
streamlit-option-menu==0.3.6

# Configuration and Logging
python-dotenv==1.0.0
loguru==0.7.0
pyyaml==6.0.1

# Utilities
joblib==1.3.1
tqdm==4.65.0
click==8.1.6

# Testing
pytest==7.4.0
pytest-cov==4.1.0
pytest-mock==3.11.1

# Code Quality
black==23.7.0
flake8==6.1.0
mypy==1.4.1
isort==5.12.0
pre-commit==3.3.3

# Deployment
gunicorn==21.2.0
fastapi==0.100.0
uvicorn==0.23.1

# Optional for GPU support
# tensorflow==2.13.0
# torch==2.0.1